avoid meta data requests in directory creation.
This is related to a bad slow down on BlueWaters. mkdir has to be
called on every rank to ensure the file system is propagated.
But if we try to verify this with stat it because really slow.

Therefore we shall only use stat to check on known mkdir failures.
if mkdir fails with anything other than EEXIST we revert to use
stat to test the existance.

We have to do this because I don't think lusture guarentees ordering
of access. AKA if one client creates a file it is not immediately visible
to another client?

Otherwise we can use a single rank to create the directory and use
a MPI Barrier.