remove the checksum API.
Currently if a data set is modified after first write, checksum will
likely be violated. This can be a feature not a bug.

To purge a dataset one need to open a BigBlock with the intention
to recreate. This can currently be done with two function calls,
first to open, copy over the meta data, then call create.