Use a smarter aggregation algorithm.
This will automatically bundle data into segments of roughly
the AggregationThreshold in size, then group the segments
into writer groups, and throttle in the group.

should have helped in reducing the number of throttled IOs when
each rank has data that is smaller than the total.

Also should have improved the load balancing as segments are of the same size.